Fix compile on Solaris. Thanks for notyfiyng, sadly is too late for KDE 3.4.0, will come with KDE 3.4.1 BUGS: 100778
svn path=/branches/KDE_3_4_BRANCH/kdegraphics/kpdf/; revision=394902

+++ ui/pageview.cpp
@@ -1497,7 +1497,7 @@ void PageView::slotRelayoutPages()
         // Here we find out column's width and row's height to compute a table
         // so we can place widgets 'centered in virtual cells'.
         int nCols = Settings::viewColumns(),
-            nRows = (int)ceilf( (float)pageCount / (float)nCols ),
+            nRows = (int)ceil( (float)pageCount / (float)nCols ),
             * colWidth = new int[ nCols ],
             * rowHeight = new int[ nRows ],
             cIdx = 0,
